Introduce margin parameter to ZenRoundedContainer
Added a 'margin' parameter to the ZenRoundedContainer ViewModifier enabling further customization of the view's surrounding space. This adjustment allows for better control over the view's layout, making it possible to define both the padding inside the container and the margin outside of it, increasing layout flexibility and improving visual consistency when the view is used in various contexts.